+++ crates/swc_ecma_lints/src/rules/radix.rs
@@ -1,14 +1,9 @@
-use std::{
-    fmt::{self, Debug},
-    sync::Arc,
-};
-
 use serde::{Deserialize, Serialize};
 use swc_atoms::JsWord;
 use swc_common::{
     collections::AHashSet,
     errors::{DiagnosticBuilder, HANDLER},
-    SourceMap, Span, SyntaxContext,
+    Span, SyntaxContext,
 };
 use swc_ecma_ast::*;
 use swc_ecma_utils::{collect_decls_with_ctxt, ident::IdentLike};
@@ -51,14 +46,12 @@ pub struct RadixConfig {
 
 pub fn radix(
     program: &Program,
-    source_map: &Arc<SourceMap>,
     top_level_ctxt: SyntaxContext,
     config: &RuleConfig<RadixConfig>,
 ) -> Option<Box<dyn Rule>> {
     match config.get_rule_reaction() {
         LintRuleReaction::Off => None,
         _ => Some(visitor_rule(Radix::new(
-            source_map.clone(),
             collect_decls_with_ctxt(program, top_level_ctxt),
             top_level_ctxt,
             config,
@@ -66,12 +59,11 @@ pub fn radix(
     }
 }
 
-#[derive(Default)]
+#[derive(Default, Debug)]
 struct Radix {
     expected_reaction: LintRuleReaction,
     top_level_ctxt: SyntaxContext,
     top_level_declared_vars: AHashSet<Id>,
-    source_map: Arc<SourceMap>,
 
     radix_mode: RadixMode,
     check_parens: bool,
@@ -79,31 +71,10 @@ struct Radix {
     classes_depth: usize,
     objects_depth: usize,
     arrow_fns_depth: usize,
-    obj: Option<JsWord>,
-    prop: Option<JsWord>,
-}
-
-impl Debug for Radix {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("Radix")
-            .field("expected_reaction", &self.expected_reaction)
-            .field("top_level_ctxt", &self.top_level_ctxt)
-            .field("top_level_declared_vars", &self.top_level_declared_vars)
-            .field("radix_mode", &self.radix_mode)
-            .field("check_parens", &self.check_parens)
-            .field("classes_depth", &self.classes_depth)
-            .field("classes_depth", &self.classes_depth)
-            .field("objects_depth", &self.objects_depth)
-            .field("arrow_fns_depth", &self.arrow_fns_depth)
-            .field("obj", &self.obj)
-            .field("prop", &self.prop)
-            .finish()
-    }
 }
 
 impl Radix {
     fn new(
-        source_map: Arc<SourceMap>,
         top_level_declared_vars: AHashSet<Id>,
         top_level_ctxt: SyntaxContext,
         config: &RuleConfig<RadixConfig>,
@@ -114,7 +85,6 @@ impl Radix {
             expected_reaction: config.get_rule_reaction(),
             top_level_ctxt,
             top_level_declared_vars,
-            source_map,
 
             radix_mode: rule_config.mode.unwrap_or_default(),
             check_parens: rule_config.check_parens.unwrap_or(true),
@@ -122,8 +92,6 @@ impl Radix {
             classes_depth: 0,
             objects_depth: 0,
             arrow_fns_depth: 0,
-            obj: None,
-            prop: None,
         }
     }
 
@@ -185,7 +153,6 @@ impl Radix {
         match call_expr.args.get(1) {
             Some(ExprOrSpread { expr, .. }) => {
                 match &extract_arg_val(
-                    &self.source_map,
                     &self.top_level_ctxt,
                     &self.top_level_declared_vars,
                     expr.as_ref(),